requirements.txt
pywin32
olefile
Aspose.Email-for-Python-via-NET